rename message package to util
