+++ src/main/java/io/github/kjm015/kylenewer/commands/StoryCommand.kt
@@ -1,0 +1,18 @@
+package io.github.kjm015.kylenewer.commands
+
+import com.jagrosh.jdautilities.command.Command
+import com.jagrosh.jdautilities.command.CommandEvent
+
+class StoryCommand : Command() {
+
+    init {
+        this.name = "story"
+        this.aliases = arrayOf("tale", "fable")
+        this.help = "Kyle will tell you a story about his life."
+    }
+
+    override fun execute(event: CommandEvent?) {
+
+    }
+
+}
