+++ src/main/java/io/github/kjm015/kylenewer/util/QuotesGenerator.kt
@@ -1,4 +1,4 @@
-package io.github.kjm015.kylenewer.message
+package io.github.kjm015.kylenewer.util
 
 import java.util.*
