+++ src/test/java/io/github/kjm015/kylenewer/commands/FetchCommandTests.java
@@ -1,7 +1,6 @@
 package io.github.kjm015.kylenewer.commands;
 
 import com.jagrosh.jdautilities.command.CommandEvent;
-import io.github.kjm015.kylenewer.message.MessageGenerator;
 import net.dv8tion.jda.core.JDA;
 import net.dv8tion.jda.core.entities.User;
 import org.junit.Test;
@@ -9,7 +8,6 @@ import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.mockito.Spy;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import java.util.ArrayList;
